+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Releases
+## 0.3.0 - Protocols Configuration Revamp
+The library now requires you to describe the format of the configuration object required by your protocol. Use the "meta-protocol.json" file for this, with a new key: `"configurationFormat"`, and the value must follow the Object Definition format.
+
+This also means you no longer need to define the validation function yourself, as we use the Object Definition Library to do this validation.
+
 ## 0.2.7 - Fixes execution of meta-protocol-check
 There was a typo on the `package.json` file, preventing it to be executed properly.
 ## 0.2.6 - Update DB protocols Inteface